Feat: Task 프론트와 연결

+++ src/main/java/com/nl/sprinterbe/domain/backlog/dto/BacklogTaskCompleteRateResponse.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class BacklogTaskCompleteRateResponse {
-    private Long completeRate;
+    private int completeRate;
 }
